thead: Add GPU, DPU, HDMI nodes for testing
In this commit I've reverted the clock implementation from the new
upstream one to the old - from vendor. This is due to difficulties in
programming the PLL clocks for the DPU, in my testing there were
problems in setting the pixclk. I haven't quite figured the bug yet in
the upstream implementation, so for now I simply reverted to the vendor
proven to work implementation.

Signed-off-by: Michal Wilczynski <m.wilczynski@samsung.com>