drm/verisilicon: Add Innosilicon HDMI support for StarFive JH7110
Add a driver for the StarFive JH7110's specific implementation of the
Innosilicon HDMI transmitter bridge.

This driver acts as a component that binds to the main Verisilicon
DC8200 driver. It provides the SoC-specific callbacks required by the
generic Innosilicon HDMI bridge driver, which is exposed via the
inno_hdmi_bind() function.

The primary responsibility of this driver is to handle the complex PLL
and PHY configuration required to generate the correct TMDS clock for a
given video mode on the JH7110. It uses a set of predefined tables to
tune the PLL for various standard display timings.

An encoder is created to represent the HDMI output, which is then
attached to the generic HDMI bridge and a connector.

Signed-off-by: Michal Wilczynski <m.wilczynski@samsung.com>