dt-bindings: display: Add bindings for Verisilicon DC8200
Add a Device Tree binding for the Verisilicon DC8200 display controller.
This IP is found on multiple SoCs, including the StarFive JH7110 and the
T-HEAD TH1520.

The binding is structured generically to accommodate platform specific
differences in the IP's integration. For example, the T-HEAD TH1520
implementation uses a different number of clocks and does not have a
'resets' property, unlike the StarFive JH7110.

To handle these variations, the binding defines a common set of possible
properties, and then uses conditional 'allOf' blocks to enforce the
specific requirements for each SoC's compatible string. This patch adds
the complete binding definition for the StarFive JH7110.

Signed-off-by: Michal Wilczynski <m.wilczynski@samsung.com>